권한 체크 코드 이동
