+++ app/src/main/java/com/example/receiptcareapp/fragment/CameraFragment.kt
@@ -107,6 +107,4 @@ class CameraFragment : Fragment() {
     }
 
 
-
-
 }
